requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
tensorflow==2.15.0
numpy==1.26.2
pygame==2.5.2
pymunk==6.6.0
pillow==10.1.0
matplotlib>=3.2.2
opencv-python>=4.1.2
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=2.2
tqdm>=4.41.0
sklearn


